code block minigame validation is fully operational.
