+++ src/index.html
@@ -88,7 +88,6 @@
                     </pre>
                 </div>
                 <textarea class="textarea" id="playerInput" placeholder="Type the code!"></textarea>
-
             </section>
             <footer class="modal-card-foot">
                 <button id="submitCodeBlock" class="button">Submit</button>
@@ -96,7 +95,6 @@
             </footer>
         </div>
     </div>
-
 
     <footer></footer>
