+++ src/scripts/CodeBlocks/submitCodeBlock.js
@@ -10,17 +10,23 @@ const submitCodeBlock = (complication) => {
 
 	if (complication === 'correct'){
 		PC.confidence += 0.5
+		PC.workPoints += 20
 	} else if (complication === 'timid'){
 		PC.confidence += 1.5
+		PC.workPoints += 30
 	} else if (complication === 'delusional'){
 		PC.confidence += -1.5
+		PC.workPoints += 25
 	} else {
 		console.log(`What the frak? You passed ${complication} as the code complication from the CodeBlocks/submitCodeBlock module.`)
 	}
 
 	dbSave(db)
 	updateAllBars()
+	$('#playerInput').val('')
 	$('#modalCanvas').removeClass('is-active')
+	const buildStartCodeBlockButton = require('./buildStartCodeBlockButton')
+	buildStartCodeBlockButton()
 	startTime()
 }
