+++ src/scripts/Database/cb_initializeCodeBlocks.js
@@ -10,7 +10,7 @@ const initializeCodeBlocks = () => {
 	create_CodeBlock('JavaScript', 'Declaring Functions', 0, 'function myFunction (arg1, arg2) {\n var str = arg1\n var newStr = arg1 + arg2\n return newStr\n}', '________ myFunction (arg1, arg2) {\n ___ str _ arg1\n var newStr = ____ + arg2\n return ___Str\n}', 'function myFunction (arg1 arg2) {\n var str = arrg1\n newStr = arg1 += arg2\n reurn newStr\n}', 'function myFunction (arg1, arg2) {Æ var str = arg1Æ var newStr = arg1 + arg2Æ return newStrÆ}')
 	create_CodeBlock('HTML', 'HTML Page Strucure', 0, '<html>\n <head>\n\n </head>\n <body>\n\n </body>\n</html>', '<html>\n <____>\n\n </____>\n <body>\n\n </body>\n</____>', '<http>\n <shed>\n\n </hear>\n <bawdy>\n\n </shoddy>\n</hello>', '<html>Æ <head>ÆÆ </head>Æ <body>ÆÆ </body>Æ</html>')
 	create_CodeBlock('CSS', 'CSS Button Styling', 0, 'button {\n background-color: black;\n color: #FFF;\n}', 'button {\n background-_____: black;\n color: ____;\n}', 'button {\n body-bgcolor: #67hrs;\n color: #67hrs;\n}', 'button {Æ background-color: black;Æ color: #FFF;Æ}')
-	create_CodeBlock('JavaScript', 'Arrow Functions', 1, 'const myFunction = () => {\n const str1 = \'some text\'\n console.log\nstr1.length)\n}', 'const myFunction = __ __ {\n _____ str1 = \'some text\'\n console.___\nstr1.length)\n}', 'var myFunction = (is perfect) = {\n str1 = some text\n console.slog\nstr1.lenth)\n}', 'const myFunction = () => {Æ const str1 = \'some text\'Æ console.logÆstr1.length)Æ}')
+	create_CodeBlock('JavaScript', 'Arrow Functions', 1, 'const myFunction = () => {\n const str1 = \'some text\'\n console.log\nstr1.length)\n}', 'const myFunction = __ __ {\n _____ str1 = \'some text\'\n console.___\nstr1.length)\n}', 'var myFunction = (is perfect) = {\n str1 = some text\n console.slog\nstr1.lenth)\n}', 'const myFunction = () => {Æ const str1 = \'some text\'Æ console.log(str1.length)Æ}')
 	create_CodeBlock('Command Line', 'Command Line Alias', 1, 'alias gpom="git push origin master"', 'alias gpom="git push ______ master"', 'alies gpom="git push origin faster"', 'alias gpom="git push origin master"')
 	create_CodeBlock('Command Line', 'Command Line Function', 1, 'copyFiles() {\n  cp -r ~/workspace/ ~/workspace/backup\n}', 'copyFiles__ {\n  cp __ ~/workspace/ ~/workspace/backup\n}', 'copyFlies() {\n  copy -r ~/workspace/ /workspace/backup\n}', 'copyFiles() {Æ    cp -r ~/workspace/ ~/workspace/backupÆ}')
 	create_CodeBlock('Command Line', 'Command Line', 1, 'mkdir NewDirectory', 'mk___ NewDirectory', 'makedir NewDirectory', 'mkdir NewDirectory')
