+++ src/scripts/Time/tick.js
@@ -2,6 +2,9 @@ const dbLoad = require('../Helpers/dbLoader')
 const deathCheck = require('./deathCheck')
 const updateAllBars = require('../DOM/updateAllBars')
 const dbSave = require('../Helpers/dbSaver')
+const addHistory = require('../DOM/addHistory')
+
+let tickCounter = 0
 
 const tick = () => {
 	const db = dbLoad()
@@ -20,8 +23,10 @@ const tick = () => {
 			PC.confidence += confidenceDecay
 			dbSave(db)
 			updateAllBars()
+			addHistory(`tick ${tickCounter}`)
 		}
 	}
+	tickCounter++
 }
 
 module.exports = tick
